+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Restored `PybricksHub.output` attribute ([support#1037]).
+
+[support#1037]: https://github.com/orgs/pybricks/discussions/1037
+
 ## [1.0.0-alpha.43] - 2023-04-19
 
 ### Added